WAT/requirements.txt
python-dotenv
requests
firecrawl-py
